==> 2. FIFA Event Analysis.ipynb <==
# PROJECT TITLE: FINDING THE MOST INTERESTING TWEETS AND FAN SENTIMENT DURING A GAME TELECAST IN REALTIME

# PART 2: ANALYSIS OF SCHEDULE & EVENT DATA:

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

## DATASET CONTENT:

knockouts.csv has the following columns:
1. Category - Match Category in Knockout Tournament
2. DateTime - Start of Match Date and Time (Russia)
3. Team1 - First Team
4. Team2 - Second Team
5. Score1 - No of Goals Scored by First Team
6. Score2 - No of Goals Scored by Second Team
7. PenScore1 - No of Goals Scored by First Team in Penalties
8. PenScore2 - No of Goals Scored by Second Team in Penalties
9. Winner - Team that won

df = pd.read_csv('knockouts.csv')
df['PenScore1'] = df['PenScore1'].fillna(0)
df['PenScore2'] = df['PenScore2'].fillna(0)
df

## 1. Get Total Count of Goals:

gc1 = 0
for i in range(16):
    gc1 += df['Score1'][i]
    gc1 += df['Score2'][i]
    gc1 += df['PenScore1'][i]
    gc1 += df['PenScore2'][i]
print('Total Goals in Knockouts : ', gc1)

## 2. Check if Winners are Valid:

for i in range(16):
    if(df['Score1'][i] + df['PenScore1'][i] > df['Score2'][i] + df['PenScore2'][i]):
        if(df['Winner'][i] != df['Team1'][i]):
            print('Error')
    else:
        if(df['Winner'][i] != df['Team2'][i]):
            print('Error')

## 3. Check Distribution of Matches Played by Teams:

team_list = []
for i in range(16):
    team_list.append(df['Team1'][i])
    team_list.append(df['Team2'][i])

team_set = []
for i in team_list:
    if i not in team_set:
        team_set.append(i)

for i in team_set:
    count = team_list.count(i)
    print("Team : ", i)
    print("Match Count : ", count)
    if (count<1 or count>4):
        print("Error")
    print("\n")

# DATASET CONTENT:

events.csv has the following columns:
1. Timestamp - Indicates Time in Match
2. ExtraMins - Indicates Overtime Given in Half or Extra Time
3. Team1 - First Team
4. Team2 - Second Team
5. Player - Player Name
6. Team - Team Name of Player
7. Event - Event Occured at Given Timestamp by given player

df1 = pd.read_csv('events.csv')
df1

## 4. Check Distribution of Events of Teams:

event_list = []
for i in range(1465):
    event_list.append(df1['Team'][i])

event_set = []
for i in event_list:
    if i not in event_set:
        event_set.append(i)

event_set = [x for x in event_set if str(x) != 'nan']
for i in event_set:
    count = event_list.count(i)
    print("Team : ", i)
    print("Event Count : ", count)
    print("\n")

## 5. Check Distribution of Goals:

goal_list = []
for i in range(1465):
    if df1["Event"][i] == 'GOAL':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        goal_list.append(tmp)

goal_set = []
for i in goal_list:
    if i not in goal_set:
        goal_set.append(i)

gc2 = 0
for i in goal_set:
    count = goal_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Goal Count : ", count)
    gc2 += count
    print("\n")

## 6. Check Distribution of Fouls:

foul_list = []
for i in range(1465):
    if df1["Event"][i] == 'FOUL':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        foul_list.append(tmp)

foul_set = []
for i in foul_list:
    if i not in foul_set:
        foul_set.append(i)

for i in foul_set:
    count = foul_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Foul Count : ", count)
    print("\n")

## 7. Check Distribution of Free Kicks:

fk_list = []
for i in range(1465):
    if df1["Event"][i] == 'FREE KICK':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        fk_list.append(tmp)

fk_set = []
for i in fk_list:
    if i not in fk_set:
        fk_set.append(i)

for i in fk_set:
    count = fk_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Free Kick Count : ", count)
    print("\n")

## 8. Check Distribution of Attempts:

attempt_list = []
for i in range(1465):
    if df1["Event"][i] == 'ATTEMPT GOAL':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        attempt_list.append(tmp)

attempt_set = []
for i in attempt_list:
    if i not in attempt_set:
        attempt_set.append(i)

for i in attempt_set:
    count = attempt_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Attempt Count : ", count)
    print("\n")

## 9. Check Distribution of Blocks:

block_list = []
for i in range(1465):
    if df1["Event"][i] == 'BLOCK':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        block_list.append(tmp)

block_set = []
for i in block_list:
    if i not in block_set:
        block_set.append(i)

for i in block_set:
    count = block_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Block Count : ", count)
    print("\n")

## 10. Check Distribution of Corners:

corner_list = []
for i in range(1465):
    if df1["Event"][i] == 'CORNER':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        corner_list.append(tmp)

corner_set = []
for i in corner_list:
    if i not in corner_set:
        corner_set.append(i)

for i in corner_set:
    count = corner_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Corner Count : ", count)
    print("\n")

## 11. Check Distribution of Offsides:

offside_list = []
for i in range(1465):
    if df1["Event"][i] == 'OFFSIDE':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        offside_list.append(tmp)

offside_set = []
for i in offside_list:
    if i not in offside_set:
        offside_set.append(i)

for i in offside_set:
    count = offside_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Offside Count : ", count)
    print("\n")

## 12. Check Distribution of Cards:

yc_list = []
for i in range(1465):
    if df1["Event"][i] == 'YELLOW CARD':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        yc_list.append(tmp)

yc_set = []
for i in yc_list:
    if i not in yc_set:
        yc_set.append(i)

for i in yc_set:
    count = yc_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Yellow Card Count : ", count)
    print("\n")

rc_list = []
for i in range(1465):
    if df1["Event"][i] == 'RED CARD':
        tmp = []
        tmp.append(df1['Player'][i])
        tmp.append(df1['Team'][i])
        rc_list.append(tmp)

rc_set = []
for i in rc_list:
    if i not in rc_set:
        rc_set.append(i)

for i in rc_set:
    count = rc_list.count(i)
    print("Player : ", i[0])
    print("Team : ", i[1])
    print("Red Card Count : ", count)
    print("\n")

## 13. Check Match List:

match_list = []
for i in range(1465):
    tmp = []
    tmp.append(df1['Team1'][i])
    tmp.append(df1['Team2'][i])
    match_list.append(tmp)

match_set = []
for i in match_list:
    if i not in match_set:
        match_set.append(i)

match_set               

## 14. Validate Goal Count:

if gc1 == gc2:
    print('Valid Goal Count')
