==> 3. Fifa Preproc Ev Data.ipynb <==
# PROJECT TITLE: FINDING THE MOST INTERESTING TWEETS AND FAN SENTIMENT DURING A GAME TELECAST IN REALTIME

# PART 3: PREPROCESSING OF SCHEDULE & EVENT DATA:

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('knockouts.csv')
df['PenScore1'] = df['PenScore1'].fillna(0)
df['PenScore2'] = df['PenScore2'].fillna(0)
df

df1 = pd.read_csv('events.csv')
df1

# 1. Set Event IDs

df1['MatchNo'] = 0
cnt = 0
for i in range(1466):
    if df1['Timestamp'][i] == 0:
        cnt += 1
    df1['MatchNo'][i] = cnt

df1['SPNo'] = 0
for i in range(1466):
    if df1['Timestamp'][i] == 0 and df1['Event'][i] == 'START':
        cnt = 1
    if df1['Timestamp'][i] == 45 and df1['Event'][i] == 'START':
        cnt += 1
    if df1['Timestamp'][i] == 90 and df1['Event'][i] == 'START':
        cnt += 1
    if df1['Timestamp'][i] == 105 and df1['Event'][i] == 'START':
        cnt += 1
    if df1['Timestamp'][i] == 120 and df1['Event'][i] == 'START':
        cnt += 1
    df1['SPNo'][i] = cnt

df1['EvNo'] = 0
for i in range(1466):
    if df1['Timestamp'][i] == 0 and df1['Event'][i] == 'START':
        cnt = 1
    if df1['Timestamp'][i] == 45 and df1['Event'][i] == 'START':
        cnt = 1
    if df1['Timestamp'][i] == 90 and df1['Event'][i] == 'START':
        cnt = 1
    if df1['Timestamp'][i] == 105 and df1['Event'][i] == 'START':
        cnt = 1
    if df1['Timestamp'][i] == 120 and df1['Event'][i] == 'START':
        cnt = 1
    df1['EvNo'][i] = cnt
    cnt += 1

df1['EventID'] = 0
tmp = 0
for i in range(1466):
    df1['EventID'][i] = (df1['MatchNo'][i] * 1000) + (df1['SPNo'][i] * 100) + df1['EvNo'][i]
df1

# 2. Set Event Time

df2 = df['DateTime']
time_array = []
for i in df2:
    x = i.split(" ")
    time_array.append(x[1])
time_array

ta2 = []
hr = 0
for i in time_array:
    y = i.split(":")
    hr = int(y[0])
    ta2.append(hr)
ta2

df1['EventHr'] = 0
df1['EventMnt'] = 0
df1['TSEM'] = 0
for i in range(1466):
    df1['EventHr'][i] = ta2[df1['MatchNo'][i] - 1]
    df1['TSEM'][i] = df1['Timestamp'][i] + df1['ExtraMins'][i]
df1

def time_calc(mn, tsem_arr):
    a = []
    

    for i in range(1466):
        if df1['MatchNo'][i] == mn:
            a.append(df1['TSEM'][i])
    
    for j in range(4):
        diff = 0
        pos = 0
        for i in range(len(a)):
            if a[i] < a[i-1] and a[i] != 0:
                diff = a[i-1] - a[i]
                pos = i
                break

        for i in range(len(a)):
            if i >= pos:
                a[i] += diff
            
    tsem_arr.append(a)
    return tsem_arr

tsem_arr = []
tsem_proc = []
for i in range(1, 17):
    tsem_arr = time_calc(i, tsem_arr)
for i in tsem_arr:
    for j in i:
        tsem_proc.append(j)
df1['TSEM_Proc'] = tsem_proc
df1

for i in range(1466):
    df1['EventMnt'][i] = df1['TSEM_Proc'][i]
    if df1['EventMnt'][i] > 59 and df1['EventMnt'][i] < 120:
        df1['EventMnt'][i] -= 60
        df1['EventHr'][i] += 1
    if df1['EventMnt'][i] > 119:
        df1['EventMnt'][i] -= 120
        df1['EventHr'][i] += 2

df1['Time'] = ''
for i in range(1466):
    if df1['EventMnt'][i] > 9:
        df1['Time'][i] = str(df1['EventHr'][i]) + ':' + str(df1['EventMnt'][i])
    else:
        df1['Time'][i] = str(df1['EventHr'][i]) + ':0' + str(df1['EventMnt'][i])

df1

# 3. Set Event Date

df2 = df['DateTime']
date_array = []
for i in df2:
    x = i.split(" ")
    date_array.append(x[0])
date_array

df1['Date'] = ''
for i in range(1466):
    df1['Date'][i] = date_array[df1['MatchNo'][i] - 1]

df1.drop(['SPNo', 'EvNo', 'TSEM', 'TSEM_Proc'], axis = 1, inplace = True)
df1

df1.to_csv('events_proc.csv')
