==> 4. Fifa Preproc Tweet Data.ipynb <==
# PROJECT TITLE: FINDING THE MOST INTERESTING TWEETS AND FAN SENTIMENT DURING A GAME TELECAST IN REALTIME

# PART 4: PREPROCESSING OF TWEET DATA:

import numpy as np
import pandas as pd
import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('FIFA.csv')
df.info()

The Following Columns are important for sentiment analysis:
1. Date - Map Event in matches to tweets
2. Tweet - Actual Tweet Content
3. Likes - Weight of Sentiment
4. RTs - Weight of Sentiment
5. Hashtags - Hashtags often give reference to player / team
6. UserMentionedNames / IDs - IDs give reference to player / team

df.drop(['ID', 'lang', 'Source', 'len', 'Orig_Tweet', 'Name','Place', 'Followers', 'Friends'], axis = 1, inplace = True)
df

df1 = df['Date']
date_array = []
time_array = []
for i in df1:
    x = i.split(" ")
    date_array.append(x[0])
    time_array.append(x[1])
df['date'] = date_array
df['time'] = time_array
df.drop(['Date'], axis = 1, inplace = True)
df

df1 = df['Likes']
df2 = df['RTs']
like_array = []
rt_array = []
weight_array = []

for i in df1:
    like_array.append(i)
for i in df2:
    rt_array.append(i)

for i in range(len(like_array)):
    weight_array.append(like_array[i] + rt_array[i])

df['weight'] = weight_array
df.drop(['Likes', 'RTs'], axis = 1, inplace = True)
df

ht_arr = []
df.Hashtags = df.Hashtags.fillna('')
df1 = df['Hashtags']
for i in df1:
    x = i.split(',')
    ht_arr.append(x)
df['HTs'] = ht_arr
df.drop(['Hashtags'], axis = 1, inplace = True)
df

umn_arr = []
df.UserMentionNames = df.UserMentionNames.fillna('')
df1 = df['UserMentionNames']
for i in df1:
    x = i.split(',')
    umn_arr.append(x)
df['UMN'] = umn_arr
df.drop(['UserMentionNames'], axis = 1, inplace = True)
df

umid_arr = []
df.UserMentionID = df.UserMentionID.fillna('')
df1 = df['UserMentionID']
for i in df1:
    x = i.split(',')
    umid_arr.append(x)
df['UMID'] = umid_arr
df.drop(['UserMentionID'], axis = 1, inplace = True)
df

tmp = []
for i in range(len(ht_arr)):
    tmp2 = []
    for j in ht_arr[i]:
        tmp2.append(j.lower())
    for k in umn_arr[i]:
        tmp2.append(k.lower())
    for l in umid_arr[i]:
        tmp2.append(l.lower())
    tmp.append(tmp2)
df['Tweet_Extra'] = tmp
df

day_arr = []
mon_arr = []
df1 = df['date']
for i in df1:
    x = i.split('-')
    mon_arr.append(int(x[1]))
    day_arr.append(int(x[2]))
df['day'] = day_arr
df['mon'] = mon_arr
df.drop(['date'], axis = 1, inplace = True)
df

hr_arr = []
mnt_arr = []
df1 = df['time']
for i in df1:
    x = i.split(':')
    hr_arr.append(int(x[0]))
    mnt_arr.append(int(x[1]))
df['hr'] = hr_arr
df['mnt'] = mnt_arr
df.drop(['time'], axis = 1, inplace = True)
df

df1 = df.sort_values(by = ['mon','day','hr','mnt'])
df1.reset_index(inplace = True, drop = True)
df1

df1.to_csv('tweets_proc.csv')
