==> 7. Fifa Agg Sent.ipynb <==
# PROJECT TITLE: FINDING THE MOST INTERESTING TWEETS AND FAN SENTIMENT DURING A GAME TELECAST IN REALTIME

# PART 7: GET AGGREGATE FAN SENTIMENT OF TEAMS:

import numpy as np
import pandas as pd
import warnings
warnings.filterwarnings('ignore')

def getScore(tn, team_weight):
    team_score = 0
    for i in team_weight:
        if tn in i[0]:
            team_score += i[1]
    return team_score

# Team List
team_list = []
# Team Set
team_set = []
dfEv = pd.read_csv('events_proc.csv')
dfEv.Player = dfEv.Player.fillna('')
dfEv.Team = dfEv.Team.fillna('')

for i in range(1466):
    if dfEv['Player'][i] != '':
        team_list.append(dfEv['Team'][i].lower())

for i in team_list:
    if i not in team_set:
        team_set.append(i)

def getTeamSent(fn):
    # Team Reffered To List
    team_ref = []
    # Team Reffered To Set
    team_ref_set = []
    # Weighted Sentiment
    ws = []
    # Team Weight = Team Ref + WS
    team_weight = []
    # All teams Sentiments
    all_team_sent = []
                
    df = pd.read_csv(fn)
    df.drop(['Unnamed: 0', 'Tweet', 'Tweet_Extra', 'Day', 'Mon', 'Hr', 'Mnt', 'Tweet_Proc'], axis = 1, inplace = True)
    df['SentDiff'] = 0
    df['WeightedSent'] = 0
    
    for i in range(df.shape[0]):
        df['SentDiff'][i] = df['PosSent'][i] - df['NegSent'][i]
        df['WeightedSent'][i] = df['weight'][i] * df['SentDiff'][i]

    df.drop(['weight', 'PosSent', 'NegSent', 'NeuSent', 'SentDiff'], axis = 1, inplace = True)
    
    df1 = df['TeamsRef']
    for i in df1:
        x = i.split('[')
        y = x[1].split(']')
        z = y[0].split("'")
        team_ref.append(z)

    for i in team_ref:
            i.remove('')

    for i in team_ref:
        for j in i:
            if any(c.isalpha() for c in j):
                pass
            else:
                i.remove(j)

    for i in team_ref:
        tmp = []
        for j in i:
            if j not in tmp:
                tmp.append(j)
        team_ref_set.append(tmp)
        
    df1 = df['WeightedSent']
    for i in df1:
        ws.append(i)
        
    for i in range(len(ws)):
        tmp = []
        tmp.append(team_ref_set[i])
        tmp.append(ws[i])
        team_weight.append(tmp)
        
    for i in range(len(team_set)):
        tmp = []
        team_score = getScore(team_set[i], team_weight)
        tmp.append(team_set[i])
        tmp.append(team_score)
        all_team_sent.append(tmp)
    
    ats_max = 0
    ats_min = 0
    for i in all_team_sent:
        if i[1] > ats_max:
            ats_max = i[1]
        if i[1] < ats_min:
            ats_min = i[1]
            
    for i in all_team_sent:
        i[1] = np.interp(i[1], [ats_min, ats_max], [-1, 1])
        
    return all_team_sent

ats30 = getTeamSent('tweet_new_3006.csv')
ats30

ats01 = getTeamSent('tweet_new_0107.csv')
ats01

ats02 = getTeamSent('tweet_new_0207.csv')
ats02

ats04 = getTeamSent('tweet_new_0407.csv')
ats04

ats10 = getTeamSent('tweet_new_1007.csv')
ats10

ats11 = getTeamSent('tweet_new_1107.csv')
ats11

ats12 = getTeamSent('tweet_new_1207.csv')
ats12

ats15 = getTeamSent('tweet_new_1507.csv')
ats15

ats16 = getTeamSent('tweet_new_1607.csv')
ats16

avg_sent = []
for i in range(len(ats16)):
    av_st = (ats30[i][1] + ats01[i][1] + ats02[i][1] + ats04[i][1] + ats10[i][1] + ats11[i][1] + ats12[i][1] + ats15[i][1] + ats16[i][1]) / 9
    avg_sent.append(av_st)
    
team_overall_sent = []
for i in range(len(avg_sent)):
    tmp = []
    tmp.append(team_set[i])
    tmp.append(avg_sent[i])
    team_overall_sent.append(tmp)
    
tos_max = 0
tos_min = 0
for i in team_overall_sent:
    if i[1] > tos_max:
        tos_max = i[1]
    if i[1] < tos_min:
        tos_min = i[1]
            
for i in team_overall_sent:
    i[1] = np.interp(i[1], [tos_min, tos_max], [0, 1])

team_overall_sent
